refactor: import TokenResponse from diracx.core.models.auth
